Move merge logic
